msm: kgsl: Remove POPP
POPP constantly attempts to lower the GPU's frequency behind the governor's
back in order to save power; however, the GPU governor in use
(msm-adreno-tz) is very good at determining the GPU's load and selecting
an appropriate frequency to run the GPU at.

POPP was created long ago (perhaps when msm-adreno-tz didn't exist or
didn't work so well), so it is clearly deprecated. Remove it.

Change-Id: Ic5c298ed25c383ea9c8b90290360fc0dfb3a70af
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[kdrag0n: Forward-ported to k4.14 for Adreno 6xx GPUs and added a change
          that Sultan accidentally squashed into the ioctl wake commit]
Signed-off-by: Danny Lin <danny@kdrag0n.dev>